+++ client/src/components/CommonGenComponents/CoprimeCondition.tsx
@@ -1,14 +1,16 @@
+import Header5 from "../PageComponents/Headers/Header5.tsx";
+
 interface CoprimeConditionProps {
   conditionS: number;
 }
 
 const CoprimeCondition = ({ conditionS }: CoprimeConditionProps) => {
   return (
-    <div className="w-full text-center my-1">
+    <Header5>
       {conditionS === 1
         ? "Періоди є взаємно простими."
         : "Періоди не взаємно прості. Обчислення сум та множень не виконано."}
-    </div>
+    </Header5>
   );
 };
