+++ client/src/pages/MatrixGeneratorPage.tsx
@@ -121,18 +121,18 @@ const MatrixGeneratorPage = observer(() => {
             <PageHeader
               title="Матричний генератор"
               paragraph="
-              Це такий генератор бінарних послідовностей,
-              який використовує матриці для створення ПВП.
-              Принцип його роботи полягає у множенні початкового вектора на матрицю,
-              що дозволяє отримати новий вектор,
-              який використовується для генерації бітів послідовності.
-              Особливості матричного генератора включають
-              високий ступінь випадковості отриманих послідовностей,
-              можливість гнучкої настройки параметрів за допомогою
-              вибору матриць різної розмірності,
-              коефіцієнтів та рангу матриці,
-              а також його застосування в різних областях.
-            "
+                Це такий генератор бінарних послідовностей,
+                який використовує матриці для створення ПВП.
+                Принцип його роботи полягає у множенні початкового вектора на матрицю,
+                що дозволяє отримати новий вектор,
+                який використовується для генерації бітів послідовності.
+                Особливості матричного генератора включають
+                високий ступінь випадковості отриманих послідовностей,
+                можливість гнучкої настройки параметрів за допомогою
+                вибору матриць різної розмірності,
+                коефіцієнтів та рангу матриці,
+                а також його застосування в різних областях.
+              "
             />
             <hr className="border-purpleFirst opacity-30 mb-10" />
