+++ client/src/routes.ts
@@ -6,8 +6,8 @@ import {
   MATRIX_GEN_ROUTE,
   POLYNOMIALS_ROUTE,
   REGISTRATION_ROUTE,
-  SUM_AND_MULTI_GEN_ROUTE,
-  HAMMING_WEIGHT_ANALYSIS_ROUTE,
+  SUM_AND_PRODUCT_GEN_ROUTE,
+  REGISTER_COMPARISON,
 } from "./utils/consts.ts";
 import AdminPanel from "./pages/AdminPanel.tsx";
 import PolynomialsPage from "./pages/PolynomialsPage.tsx";
@@ -15,7 +15,7 @@ import AuthorizationPage from "./pages/AuthorizationPage.tsx";
 import LinearGeneratorPage from "./pages/LinearGeneratorPage.tsx";
 import MatrixGeneratorPage from "./pages/MatrixGeneratorPage.tsx";
 import SumAndProductGeneratorPage from "./pages/SumAndProductGeneratorPage.tsx";
-import HammingWeightAnalysisPage from "./pages/HammingWeightAnalysisPage.tsx";
+import RegisterComparisonPage from "./pages/RegisterComparisonPage.tsx";
 import MainPage from "./pages/MainPage.tsx";
 import RegistrationPage from "./pages/RegistrationPage.tsx";
 
@@ -50,13 +50,13 @@ export const publicRoutes = [
   },
   {
     name: "Sum and product register",
-    path: SUM_AND_MULTI_GEN_ROUTE,
+    path: SUM_AND_PRODUCT_GEN_ROUTE,
     Component: SumAndProductGeneratorPage,
   },
   {
-    name: "Hamming weight analysis",
-    path: HAMMING_WEIGHT_ANALYSIS_ROUTE,
-    Component: HammingWeightAnalysisPage,
+    name: "Register comparison",
+    path: REGISTER_COMPARISON,
+    Component: RegisterComparisonPage,
   },
 ];
